Fixed linearized image size error

+++ src/edu/stanford/rsl/conrad/opencl/OpenCLForwardProjector.java
@@ -349,7 +349,7 @@ public class OpenCLForwardProjector implements GUIConfigurable, Citeable {
 					memorysize = (int) (volumeSize[0] * volumeSize[1] * subVolumeZ * Float.SIZE / 8);
 					if(debug)System.out.println("Memory: " + memorysize);
 				} else {
-					h_volume = new float[(int) (volumeSize[0] * volumeSize[1] * volumeSize[2])];
+					h_volume = new float[(int)volumeSize[0] * (int)volumeSize[1] * (int)volumeSize[2]];
 					subVolumeZ = (int) volumeSize[2];
 					nSteps = 1;
 				}
